+++ src/pages/api/recurring.ts
@@ -12,6 +12,7 @@ const cors = initMiddleware(
 );
 
 const API_URL = globalVars().API_URL;
+// url: API_URL + '/api/earmark/recurring',
 
 export default async function handler(req, res) {
     // Run cors
@@ -38,6 +39,7 @@ export default async function handler(req, res) {
                 },
             };
             const { data } = await axios(config);
+            console.log('prod data', data)
             finalResponse = data;
             finalStatus = 200;
         } catch (error) {